refactor: move copyColorData
