+++ packages/runtime/src/controls/style.ts
@@ -1,10 +1,9 @@
-import { ColorData, ResponsiveValue } from '@makeswift/controls'
+import { ColorData, copyColorData, ResponsiveValue } from '@makeswift/controls'
 import { BorderRadiusPropertyData } from '../css/border-radius'
 import type { LengthPercentageData } from '../css/length-percentage'
 import { MarginPropertyData } from '../css/margin'
 import { PaddingPropertyData } from '../css/padding'
 import { CopyContext } from '../state/react-page'
-import { copyColorData } from './color'
 import { Send } from '../prop-controllers/instances'
 import { PropController } from '../prop-controllers/base'
 import { BoxModel } from '../state/modules/box-models'
